순환 참조 문제 Facade 패턴 적용
