+++ src/main/java/com/grabpt/controller/MyProPageController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import com.grabpt.config.SecurityUtils;
 import com.grabpt.dto.request.*;
+import com.grabpt.service.ProfileService.ProfileFacade;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -37,7 +38,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class MyProPageController {
 
-	private final ProfileService profileService;
+	private final ProfileFacade profileFacade;
 	private final ObjectMapper objectMapper;
 	private final UserQueryService userQueryService;
 
@@ -46,7 +47,7 @@ public class MyProPageController {
 	public ApiResponse<ProfileResponseDTO.MyProProfileDTO> getMyProUserProfile()  {
 
 		Long userId = SecurityUtils.currentUserIdOrThrow();
-		return ApiResponse.onSuccess(profileService.findMyProUserProfile(userId));
+		return ApiResponse.onSuccess(profileFacade.findMyProUserProfile(userId));
 	}
 
 	@GetMapping("/reviews")
@@ -57,7 +58,7 @@ public class MyProPageController {
 
 		Long userId = SecurityUtils.currentUserIdOrThrow();
 		Pageable pageable = PageRequest.of(Math.max(page - 1, 0), size);
-		Page<MyReviewListDTO> reviews = profileService.findProReviews(userId, pageable);
+		Page<MyReviewListDTO> reviews = profileFacade.findProReviews(userId, pageable);
 		return ApiResponse.onSuccess(reviews);
 	}
 
@@ -66,7 +67,7 @@ public class MyProPageController {
 	public ApiResponse<CertificationResponseDTO> getProCertifications(
 		) {
 		Long userId = SecurityUtils.currentUserIdOrThrow();
-		CertificationResponseDTO certifications = profileService.findMyCertifications(userId);
+		CertificationResponseDTO certifications = profileFacade.findMyCertifications(userId);
 		return ApiResponse.onSuccess(certifications);
 	}
 
@@ -94,7 +95,7 @@ public class MyProPageController {
 		// JSON 문자열을 DTO 객체로 변환
 		CertificationUpdateRequestDTO request = objectMapper.readValue(requestJson, CertificationUpdateRequestDTO.class);
 
-		profileService.updateProCertifications(userId, request, newImages);
+		profileFacade.updateProCertifications(userId, request, newImages);
 		return ApiResponse.onSuccess("자격증 정보가 성공적으로 등록되었습니다.");
 	}
 
@@ -104,7 +105,7 @@ public class MyProPageController {
 		@Valid @RequestBody CenterUpdateRequestDTO request){
 
 		Long userId = SecurityUtils.currentUserIdOrThrow();
-		profileService.updateProCenter(userId, request);
+		profileFacade.updateProCenter(userId, request);
 		return ApiResponse.onSuccess("센터 정보 수정이 완료되었습니다.");
 	}
 
@@ -113,7 +114,7 @@ public class MyProPageController {
 	public ApiResponse<String> updateProDescription(
 		@Valid @RequestBody DescriptionUpdateRequestDTO request) {
 		Long userId = SecurityUtils.currentUserIdOrThrow();
-		profileService.updateProDescription(userId, request);
+		profileFacade.updateProDescription(userId, request);
 		return ApiResponse.onSuccess("전문가 소개가 수정되었습니다.");
 	}
 
@@ -129,7 +130,7 @@ public class MyProPageController {
 
 		Long userId = SecurityUtils.currentUserIdOrThrow();
 		PhotoUpdateRequestDTO updateRequest = objectMapper.readValue(requestJson, PhotoUpdateRequestDTO.class);
-		profileService.updateProPhotos(userId, updateRequest, newPhotoFiles);
+		profileFacade.updateProPhotos(userId, updateRequest, newPhotoFiles);
 
 		return ApiResponse.onSuccess("사진이 성공적으로 수정되었습니다.");
 	}
@@ -139,7 +140,7 @@ public class MyProPageController {
 	public ApiResponse<String> updateProPtPrice(
 		@Valid @RequestBody PtPriceRequest.PtPriceUpdateRequestList request) {
 		Long userId = SecurityUtils.currentUserIdOrThrow();
-		profileService.updateProPtPrice(userId, request);
+		profileFacade.updateProPtPrice(userId, request);
 		return ApiResponse.onSuccess("PT 가격 정보가 수정되었습니다.");
 	}
 
@@ -148,7 +149,7 @@ public class MyProPageController {
 	public ApiResponse<String> updateProProgram(
 		@Valid @RequestBody PtProgramUpdateRequestDTO request) {
 		Long userId = SecurityUtils.currentUserIdOrThrow();
-		profileService.updateProProgram(userId, request);
+		profileFacade.updateProProgram(userId, request);
 		return ApiResponse.onSuccess("PT 프로그램 정보가 수정되었습니다.");
 	}
 
@@ -158,7 +159,7 @@ public class MyProPageController {
 		@RequestBody @Valid ProLocationUpdateRequestDTO request) {
 
 		Long userId = SecurityUtils.currentUserIdOrThrow();
-		profileService.updateProLocation(userId, request);
+		profileFacade.updateProLocation(userId, request);
 		return ApiResponse.onSuccess("위치 정보가 성공적으로 수정되었습니다.");
 	}
 }
